+++ src/eg3dcam.c
@@ -12,12 +12,12 @@
 #include <dos.h>
 #include <memory.h>
 
-void setViewRotation(int rotX, int rotY, int rotZ) {
+void setViewRotation(int16 rotX, int16 rotY, int16 rotZ) {
     buildRotationMatrixFar(g_viewRotMatrix, -rotX, -rotY, -rotZ);
 }
 
 // ==== seg000:0x3a90 ====
-void setViewPosition(int viewX, int viewY, int viewZ) { /* Original: SetViewPos(X,Y,Z). Store viewer coordinates for 3D transforms. */
+void setViewPosition(int16 viewX, int16  viewY, int16 viewZ) { /* Original: SetViewPos(X,Y,Z). Store viewer coordinates for 3D transforms. */
     g_viewPosX = viewX;
     g_viewPosY = viewY;
     g_viewPosZ = viewZ;
