+++ src/egpic.c
@@ -23,7 +23,7 @@ void showPicFile(int handle, int pageNum);
 #ifdef BUGFIX
 void openBlitClosePic(const char *filename, int page) { /* Original chain: OpenFile + blit/decode + CloseFile. Open, blit PIC to page, then close. */
 #else
-void openBlitClosePic(const char *filename, int page, int garbage) { /* Original chain: OpenFile + blit/decode + CloseFile. Open, blit PIC to page, then close. */
+void openBlitClosePic(const char *filename, int16 page, int16 garbage) { /* Original chain: OpenFile + blit/decode + CloseFile. Open, blit PIC to page, then close. */
 #endif
     int handle = openFileWrapper(filename, 0);
     /* The PIC decoder/blitter consumes the already-open file handle. */
